+++ playwright/core/expectScreenshotFixture.ts
@@ -9,30 +9,29 @@ export const expectScreenshotFixture: PlaywrightFixture<ExpectScreenshotFixture>
 ) => {
     const expectScreenshot: ExpectScreenshotFixture = async ({
         component,
-        screenshotName,
+        nameSuffix,
         ...pageScreenshotOptions
     } = {}) => {
-        const captureScreenshot = async (theme: string) => {
-            const body = page.locator('body');
-
-            await body.evaluate((el: HTMLElement | SVGElement, newTheme: string) => {
-                el.setAttribute('class', `g-root g-root_theme_${newTheme}`);
-            }, theme);
-
+        const captureScreenshot = async () => {
             return (component || page.locator('.playwright-wrapper-test')).screenshot({
                 animations: 'disabled',
                 ...pageScreenshotOptions,
             });
         };
 
-        const nameScreenshot = testInfo.titlePath.slice(1).join(' ');
+        const nameScreenshot =
+            testInfo.titlePath.slice(1).join(' ') + (nameSuffix ? ` ${nameSuffix}` : '');
 
-        expect(await captureScreenshot('dark')).toMatchSnapshot({
-            name: `${screenshotName || nameScreenshot} dark.png`,
+        await page.emulateMedia({colorScheme: 'light'});
+
+        expect(await captureScreenshot()).toMatchSnapshot({
+            name: `${nameScreenshot} light.png`,
         });
 
-        expect(await captureScreenshot('light')).toMatchSnapshot({
-            name: `${screenshotName || nameScreenshot} light.png`,
+        await page.emulateMedia({colorScheme: 'dark'});
+
+        expect(await captureScreenshot()).toMatchSnapshot({
+            name: `${nameScreenshot} dark.png`,
         });
     };